Modify Task class to be abstract and define abstract getDates() method
Task class does not have any methods to retrieve dates from child
classes.

Storage Class needs the date attribute in child classes like Deadline
and Event for the saving and loading tasks correctly.

Modify Task to be an abstract class with an abstract method getDates().
getDates() can be defined in the respective child classes to return the
number of dates required for the class.

Using abstract methods in this case is preferable as to not introduce
unnecessary methods in classes that do not need them. Each child
class returns different number of dates, and this approach ensures
flexibility.