Add compareDate() method and TaskComparator class for comparing Tasks
Tasks cannot be compared meaningfully.

With LocalDateTime being used instead of Strings, it is essential to
have a consistent way of comparing Tasks.

A new abstract method compareDate() will return the date which we can
use to compare the task. TaskComparator will rank Tasks based on that
date.

This method will ensure consistency as there are tasks having multiple
dates associated with them and some having none.