Add abstract class Command and implement Commands
All the commands and functionalities are executed through switch-case
statement in Main.

The Main function is cluttered and as more commands are introduced
it will get longer.

An abstract class Command is defined and all commands will be subclasses
of Command. Functionalities of each command will be encapsulated in
its class.

This approach allows new commands to be added in easily without
cluttering the Main class.