Refactor Ella to use Command, Ui and Parser class
The logic for parsing, displaying messages to user as well as executing
commands are found in Ella.

Ella is very long and cluttered and does multiple tasks.

Separate the various components into various classes. Ui handles
user input related tasks, Parser will parse the user input and Command
will execute the functionalities of a given command.

This ensures code will be more organized as well as easier to read.